+++ src/main/resources/templates/fragment/sidebar.html
@@ -104,14 +104,6 @@
                 </div>
               </a>
             </li>
-            <li class="nav-item">
-              <a class="nav-link" href="/community">
-                <div class="d-flex align-items-center">
-                  <span class="nav-link-icon"> <i class="bi bi-people-fill"></i></span>
-                  <span class="nav-link-text">커뮤니티 관리</span>
-                </div>
-              </a>
-            </li>
           </ul>
         </div>
       </div>
@@ -215,15 +207,6 @@
                 <div class="d-flex align-items-center">
                   <span class="nav-link-icon"> <i class="bi bi-cart"></i></span>
                   <span class="nav-link-text">상점 관리</span>
-                </div>
-              </a>
-            </li>
-
-            <li class="nav-item">
-              <a class="nav-link" href="/community">
-                <div class="d-flex align-items-center">
-                  <span class="nav-link-icon"> <i class="bi bi-people-fill"></i></span>
-                  <span class="nav-link-text">커뮤니티 관리</span>
                 </div>
               </a>
             </li>
